Draw CircuitComponents and QubitComponents to the screen on Load. Some move on the screen, some don't.

+++ QMat_Calculator/Matrices/Matrix.cs
@@ -31,7 +31,7 @@ namespace QMat_Calculator.Matrices
             this.values = new Complex[rows, columns];
             this.preceder = -1;
         }
-        public Matrix(int rows, int columns, int preceder, Complex[,] data)
+        public Matrix(int rows, int columns, double preceder, Complex[,] data)
         {
             this.rows = rows;
             this.columns = columns;
